@@ -60,7 +60,7 @@ func makeRoute(n int) *api.LookupRoute {
 
 func makeRoutes(count int) api.LookupRoutes {
 	routes := make(api.LookupRoutes, 0, count)
-	for i := 0; i < count; i++ {
+	for i := range count {
 		routes = append(routes, makeRoute(i))
 	}
 	return routes
@@ -99,7 +99,7 @@ func main() {
 	backend := postgres.NewRoutesBackend(pool, cfg.Sources)
 
 	// Now insert tons of routes...
-	for i := 0; i < 10; i++ {
+	for i := range 10 {
 		routes := makeRoutes(100000)
 		t := time.Now()
 		if err := backend.SetRoutes(
