@@ -31,12 +31,9 @@ func apiPaginateRoutes(
 	totalPages := int(math.Ceil(float64(totalResults) / float64(pageSize)))
 
 	offset := page * pageSize
-	rindex := offset + pageSize
-
-	// Don't access out of bounds
-	if rindex > totalResults {
-		rindex = totalResults
-	}
+	rindex := min(
+		// Don't access out of bounds
+		offset+pageSize, totalResults)
 	if offset < 0 {
 		offset = 0
 	}
@@ -78,12 +75,9 @@ func apiPaginateLookupRoutes(
 	totalPages := int(math.Ceil(float64(totalResults) / float64(pageSize)))
 
 	offset := page * pageSize
-	rindex := offset + pageSize
-
-	// Don't access out of bounds
-	if rindex > totalResults {
-		rindex = totalResults
-	}
+	rindex := min(
+		// Don't access out of bounds
+		offset+pageSize, totalResults)
 	if offset < 0 {
 		offset = 0
 	}
