@@ -2,6 +2,7 @@ package openbgpd
 
 import (
 	"github.com/alice-lg/alice-lg/pkg/api"
+	"slices"
 )
 
 func filterReceivedRoutes(
@@ -11,11 +12,8 @@ func filterReceivedRoutes(
 	filtered := make(api.Routes, 0, len(routes))
 	for _, r := range routes {
 		received := true
-		for _, c := range rejectCommunities {
-			if r.BGP.HasLargeCommunity(c) {
-				received = false
-				break
-			}
+		if slices.ContainsFunc(rejectCommunities, r.BGP.HasLargeCommunity) {
+			received = false
 		}
 		if received {
 			filtered = append(filtered, r)
@@ -30,13 +28,7 @@ func filterRejectedRoutes(
 ) api.Routes {
 	filtered := make(api.Routes, 0, len(routes))
 	for _, r := range routes {
-		rejected := false
-		for _, c := range rejectCommunities {
-			if r.BGP.HasLargeCommunity(c) {
-				rejected = true
-				break
-			}
-		}
+		rejected := slices.ContainsFunc(rejectCommunities, r.BGP.HasLargeCommunity)
 		if rejected {
 			filtered = append(filtered, r)
 		}
